Move `keyPathToLookupExpression`
